Signs are finally working
The blockentity register doesn't actually register block entities. Rather, it registers a blockentity *type*. So in vanilla, standing signs and wall signs are registered to the blockentity type "SIGN".

The reason we can't use this blockentity type is because it is *static*. We can extend the class, but the SIGN blockentity type cannot be modified by the user.  So we create MOD_SIGN blockentitytype that extends the SIGN blockentitytype.

In addition, the standingsign *class* includes a method that will, by default, return the blockentitytype SIGN. So if we register a block of type standingsign to the blockentitytype SIGN, there will be a conflict.

That means we *also* need to create custom extensions to the block class for the four internal sign types (standing, wall, ceilinghanging, and wallhanging.